Add support to deb package  (#266)

+++ build/install/OneClickInstall/install-RedHat/install-preq.sh
@@ -79,11 +79,13 @@ if ! rpm -q mysql-community-server; then
 fi
 
 #add elasticsearch repo
+ELASTIC_VERSION="7.13.1"
+ELASTIC_DIST=$(echo $ELASTIC_VERSION | awk '{ print int($1) }')
 rpm --import https://artifacts.elastic.co/GPG-KEY-elasticsearch
 cat > /etc/yum.repos.d/elasticsearch.repo <<END
 [elasticsearch]
-name=Elasticsearch repository for 7.x packages
-baseurl=https://artifacts.elastic.co/packages/7.x/yum
+name=Elasticsearch repository for ${ELASTIC_DIST}.x packages
+baseurl=https://artifacts.elastic.co/packages/${ELASTIC_DIST}.x/yum
 gpgcheck=1
 gpgkey=https://artifacts.elastic.co/GPG-KEY-elasticsearch
 enabled=0
@@ -93,15 +95,18 @@ END
 
 #install kafka
 PRODUCT_DIR="/var/www/${product}"
-mkdir -p ${PRODUCT_DIR}/services/
-getent passwd kafka >/dev/null || useradd -m -d ${PRODUCT_DIR}/services/kafka -s /sbin/nologin -p kafka kafka
-cd ${PRODUCT_DIR}/services/kafka
-wget https://downloads.apache.org/kafka/2.7.0/kafka_2.13-2.7.0.tgz
-tar xzf kafka_*.tgz --strip 1 && rm -rf kafka_*.tgz
-chown -R kafka ${PRODUCT_DIR}/services/kafka
-cd -
+if [ "$(ls -A "$PRODUCT_DIR/services/kafka" 2> /dev/null)" == "" ]; then
+	mkdir -p ${PRODUCT_DIR}/services/
+	getent passwd kafka >/dev/null || useradd -m -d ${PRODUCT_DIR}/services/kafka -s /sbin/nologin -p kafka kafka
+	cd ${PRODUCT_DIR}/services/kafka
+	wget https://downloads.apache.org/kafka/2.7.0/kafka_2.13-2.7.0.tgz
+	tar xzf kafka_*.tgz --strip 1 && rm -rf kafka_*.tgz
+	chown -R kafka ${PRODUCT_DIR}/services/kafka
+	cd -
+fi
 
-cat > /etc/systemd/system/zookeeper.service <<END
+if [ ! -e /lib/systemd/system/zookeeper.service ]; then
+cat > /lib/systemd/system/zookeeper.service <<END
 [Unit]
 Requires=network.target remote-fs.target
 After=network.target remote-fs.target
@@ -114,8 +119,10 @@ Restart=on-abnormal
 [Install]
 WantedBy=multi-user.target
 END
+fi
 
-cat > /etc/systemd/system/kafka.service <<END
+if [ ! -e /lib/systemd/system/kafka.service ]; then
+cat > /lib/systemd/system/kafka.service <<END
 [Unit]
 Requires=zookeeper.service
 After=zookeeper.service
@@ -128,6 +135,7 @@ Restart=on-abnormal
 [Install]
 WantedBy=multi-user.target
 END
+fi
 
 # add nginx repo
 cat > /etc/yum.repos.d/nginx.repo <<END
@@ -169,14 +177,15 @@ ${package_manager} -y install epel-release \
 			make \
 			yarn \
 			dotnet-sdk-5.0 \
-			elasticsearch-7.13.1 --enablerepo=elasticsearch \
+			elasticsearch-${ELASTIC_VERSION} --enablerepo=elasticsearch \
 			mysql-server \
 			nginx \
 			supervisor \
 			postgresql \
 			postgresql-server \
 			rabbitmq-server$rabbitmq_version \
-			redis --enablerepo=remi
+			redis --enablerepo=remi \
+			java
 	
 postgresql-setup initdb	|| true
